+++ stubs/app/Orchid/Screens/PlatformScreen.php
@@ -43,19 +43,7 @@ class PlatformScreen extends Screen
      */
     public function commandBar(): iterable
     {
-        return [
-            Link::make('Website')
-                ->href('http://orchid.software')
-                ->icon('globe-alt'),
-
-            Link::make('Documentation')
-                ->href('https://orchid.software/en/docs')
-                ->icon('docs'),
-
-            Link::make('GitHub')
-                ->href('https://github.com/orchidsoftware/platform')
-                ->icon('social-github'),
-        ];
+        return [];
     }
 
     /**
